+++ client/src/interfaces/Code10.ts
@@ -1,0 +1,7 @@
+interface Code10 {
+  id: string;
+  code: string;
+  what_pages: string[];
+}
+
+export default Code10;